refactor: change deploy method
